AI_Account/requirements.txt
django==3.0.5
djangorestframework==3.12.4
djongo==1.3.6
Sentence-transformers==2.0.0
laserembeddings==1.1.1
laserembeddings[zh]
laserembeddings[ja]
langdetect==1.0.9
langid==1.1.6
pandas
opencv-python==4.5.3.56
opencv-contrib-python==4.5.3.56
imutils==0.5.4
mammoth==1.4.16
bs4==0.0.1
textblob==0.15.3
pdf2docx==0.5.1
pdfplumber==0.5.28
pikepdf==2.16.1
pdf2image==1.14.0
openpyxl==3.0.7
google-cloud-translate==3.3.2
smbprotocol==1.2.0
fasttext==0.9.1
whatlangid==1.0.11
camelot-py==0.8.2
camelot-py[cv]==0.8.2
python-bidi==0.4.2
xlsxwriter==3.0.1
fuzzywuzzy==0.18.0
pymupdf==1.18.16
xls2xlsx
